convertfloatingnumbertointeger renamed to convertfloatingtointeger
