+++ calc-components/calculators/convertfloatingtointeger.h
